mailingapi/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
Deprecated==1.2.13
Django==4.2.1
django-oauth-toolkit==2.2.0
django-rest==0.8.7
djangorestframework==3.14.0
djangorestframework-oauth==1.1.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
kombu==5.2.4
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.1
prompt-toolkit==3.0.38
psycopg2==2.9.6
pycparser==2.21
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
requests==2.30.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0


